requirements.txt
pandas
sodapy
python-dotenv
openmeteo_requests
requests_cache
retry_requests
fastapi
uvicorn
pydantic
requests
httpx
flask
folium
watchdog==3.0.0
httpxgeopy==2.4.1
